requirements.txt
openai
ollama
jinja2
google-generativeai
pyyaml
pillow
frontmatter